requirements.txt
yfinance 
pymongo 
fredapi 
newsapi-python 
python-dotenv 
pandas 
scikit-learn 
seaborn 
matplotlib 
nbformat